refactor(ollama): unify local model onboarding
