+++ extensions/ollama/src/setup.ts
@@ -1,4 +1,3 @@
-import { formatErrorMessage } from "openclaw/plugin-sdk/error-runtime";
 // Ollama setup module handles plugin onboarding behavior.
 import { expectDefined } from "openclaw/plugin-sdk/expect-runtime";
 import type {
@@ -19,7 +18,6 @@ import { applyAgentDefaultModelPrimary } from "openclaw/plugin-sdk/provider-onbo
 import type { RuntimeEnv } from "openclaw/plugin-sdk/runtime";
 import { WizardCancelledError, type WizardPrompter } from "openclaw/plugin-sdk/setup";
 import { fetchWithSsrFGuard } from "openclaw/plugin-sdk/ssrf-runtime";
-import { normalizeLowercaseStringOrEmpty } from "openclaw/plugin-sdk/string-coerce-runtime";
 import {
   OLLAMA_CLOUD_BASE_URL,
   OLLAMA_CLOUD_DEFAULT_MODELS,
@@ -30,15 +28,22 @@ import {
 import { readProviderBaseUrl } from "./provider-base-url.js";
 import {
   buildOllamaBaseUrlSsrFPolicy,
-  buildDefaultOllamaCloudModelDefinition,
   buildOllamaProvider,
-  buildOllamaModelDefinition,
   enrichOllamaModelsWithContext,
   fetchOllamaModels,
   isOllamaCloudModel,
   resolveOllamaApiBase,
   type OllamaModelWithContext,
 } from "./provider-models.js";
+import {
+  buildOllamaModelsConfig,
+  discoverOllamaModelsForSetup,
+  findAvailableOllamaModelName,
+  inspectOllamaModelsForSetup,
+  mergeUniqueModelNames,
+  normalizeOllamaModelName,
+} from "./setup-model-selection.js";
+import { pullOllamaModel, pullOllamaModelNonInteractive } from "./setup-pull.js";
 
 export { buildOllamaProvider };
 
@@ -47,13 +52,9 @@ const OLLAMA_SUGGESTED_MODELS_CLOUD = OLLAMA_CLOUD_DEFAULT_MODELS.map((model) =>
 const OLLAMA_SUGGESTED_MODELS_LOCAL_CLOUD = OLLAMA_CLOUD_DEFAULT_MODELS.map(
   (model) => `${model.id}:cloud`,
 );
-const OLLAMA_CONTEXT_ENRICH_LIMIT = 200;
 const OLLAMA_CLOUD_MAX_DISCOVERED_MODELS = 500;
-const OLLAMA_PULL_RESPONSE_TIMEOUT_MS = 30_000;
-const OLLAMA_PULL_STREAM_IDLE_TIMEOUT_MS = 300_000;
 const OLLAMA_RECOMMENDED_TOOLS_MODEL = "gemma4:e4b";
 const OLLAMA_RECOMMENDED_TOOLS_MODEL_SIZE = "about 9.6 GB";
-const OLLAMA_TOOLS_SCAN_CONCURRENCY = 8;
 
 type OllamaCloudDefaultModel = (typeof OLLAMA_CLOUD_DEFAULT_MODELS)[number];
 
@@ -113,30 +114,6 @@ function buildOllamaCloudSigninLines(signinUrl?: string): string[] {
   ];
 }
 
-function normalizeOllamaModelName(value: string | undefined): string | undefined {
-  const trimmed = value?.trim();
-  if (!trimmed) {
-    return undefined;
-  }
-  if (normalizeLowercaseStringOrEmpty(trimmed).startsWith("ollama/")) {
-    const normalized = trimmed.slice("ollama/".length).trim();
-    return normalized || undefined;
-  }
-  return trimmed;
-}
-
-function formatOllamaPullStatus(status: string): { text: string; hidePercent: boolean } {
-  const trimmed = status.trim();
-  const partStatusMatch = trimmed.match(/^([a-z-]+)\s+(?:sha256:)?[a-f0-9]{8,}$/i);
-  if (partStatusMatch) {
-    return { text: `${partStatusMatch[1]} part`, hidePercent: false };
-  }
-  if (/^verifying\b.*\bdigest\b/i.test(trimmed)) {
-    return { text: "verifying digest", hidePercent: true };
-  }
-  return { text: trimmed, hidePercent: false };
-}
-
 export async function checkOllamaCloudAuth(
   baseUrl: string,
 ): Promise<{ signedIn: boolean; signinUrl?: string }> {
@@ -170,212 +147,6 @@ export async function checkOllamaCloudAuth(
   } catch {
     return { signedIn: false };
   }
-}
-
-type OllamaPullChunk = {
-  status?: string;
-  total?: number;
-  completed?: number;
-  error?: string;
-};
-
-type OllamaPullResult = { ok: true } | { ok: false; message: string };
-
-async function readOllamaPullChunkWithIdleTimeout(
-  reader: ReadableStreamDefaultReader<Uint8Array>,
-): Promise<ReadableStreamReadResult<Uint8Array>> {
-  let timeoutId: ReturnType<typeof setTimeout> | undefined;
-  let timedOut = false;
-
-  return await new Promise((resolve, reject) => {
-    const clear = () => {
-      if (timeoutId !== undefined) {
-        clearTimeout(timeoutId);
-        timeoutId = undefined;
-      }
-    };
-
-    timeoutId = setTimeout(() => {
-      timedOut = true;
-      clear();
-      void reader.cancel().catch(() => undefined);
-      reject(
-        new Error(
-          `Ollama pull stalled: no data received for ${Math.round(OLLAMA_PULL_STREAM_IDLE_TIMEOUT_MS / 1000)}s`,
-        ),
-      );
-    }, OLLAMA_PULL_STREAM_IDLE_TIMEOUT_MS);
-
-    void reader.read().then(
-      (result) => {
-        clear();
-        if (!timedOut) {
-          resolve(result);
-        }
-      },
-      (err: unknown) => {
-        clear();
-        if (!timedOut) {
-          reject(toLintErrorObject(err, "Non-Error rejection"));
-        }
-      },
-    );
-  });
-}
-
-async function pullOllamaModelCore(params: {
-  baseUrl: string;
-  modelName: string;
-  onStatus?: (status: string, percent: number | null) => void;
-  signal?: AbortSignal;
-}): Promise<OllamaPullResult> {
-  const baseUrl = resolveOllamaApiBase(params.baseUrl);
-  const modelName = normalizeOllamaModelName(params.modelName) ?? params.modelName.trim();
-  const responseController = new AbortController();
-  const responseTimeout = setTimeout(
-    responseController.abort.bind(responseController),
-    OLLAMA_PULL_RESPONSE_TIMEOUT_MS,
-  );
-  try {
-    params.signal?.throwIfAborted();
-    const { response, release } = await fetchWithSsrFGuard({
-      url: `${baseUrl}/api/pull`,
-      init: {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ name: modelName }),
-      },
-      signal: params.signal
-        ? AbortSignal.any([responseController.signal, params.signal])
-        : responseController.signal,
-      policy: buildOllamaBaseUrlSsrFPolicy(baseUrl),
-      auditContext: "ollama-setup.pull",
-    });
-    clearTimeout(responseTimeout);
-    try {
-      if (!response.ok) {
-        return { ok: false, message: `Failed to download ${modelName} (HTTP ${response.status})` };
-      }
-      if (!response.body) {
-        return { ok: false, message: `Failed to download ${modelName} (no response body)` };
-      }
-
-      const reader = response.body.getReader();
-      const decoder = new TextDecoder();
-      let buffer = "";
-      const layers = new Map<string, { total: number; completed: number }>();
-
-      const parseLine = (line: string): OllamaPullResult => {
-        const trimmed = line.trim();
-        if (!trimmed) {
-          return { ok: true };
-        }
-        try {
-          const chunk = JSON.parse(trimmed) as OllamaPullChunk;
-          if (chunk.error) {
-            return { ok: false, message: `Download failed: ${chunk.error}` };
-          }
-          if (!chunk.status) {
-            return { ok: true };
-          }
-          if (chunk.total && chunk.completed !== undefined) {
-            layers.set(chunk.status, { total: chunk.total, completed: chunk.completed });
-            let totalSum = 0;
-            let completedSum = 0;
-            for (const layer of layers.values()) {
-              totalSum += layer.total;
-              completedSum += layer.completed;
-            }
-            params.onStatus?.(
-              chunk.status,
-              totalSum > 0 ? Math.round((completedSum / totalSum) * 100) : null,
-            );
-          } else {
-            params.onStatus?.(chunk.status, null);
-          }
-        } catch {
-          // Ignore malformed streaming lines from Ollama.
-        }
-        return { ok: true };
-      };
-
-      for (;;) {
-        const { done, value } = await readOllamaPullChunkWithIdleTimeout(reader);
-        if (done) {
-          break;
-        }
-        buffer += decoder.decode(value, { stream: true });
-        const lines = buffer.split("\n");
-        buffer = lines.pop() ?? "";
-        for (const line of lines) {
-          const parsed = parseLine(line);
-          if (!parsed.ok) {
-            return parsed;
-          }
-        }
-      }
-
-      const trailing = buffer.trim();
-      if (trailing) {
-        const parsed = parseLine(trailing);
-        if (!parsed.ok) {
-          return parsed;
-        }
-      }
-
-      return { ok: true };
-    } finally {
-      await release();
-    }
-  } catch (err) {
-    const reason = formatErrorMessage(err);
-    return { ok: false, message: `Failed to download ${modelName}: ${reason}` };
-  } finally {
-    clearTimeout(responseTimeout);
-  }
-}
-
-async function pullOllamaModel(
-  baseUrl: string,
-  modelName: string,
-  prompter: WizardPrompter,
-  signal?: AbortSignal,
-): Promise<boolean> {
-  const spinner = prompter.progress(`Downloading ${modelName}...`);
-  const result = await pullOllamaModelCore({
-    baseUrl,
-    modelName,
-    ...(signal ? { signal } : {}),
-    onStatus: (status, percent) => {
-      const displayStatus = formatOllamaPullStatus(status);
-      if (displayStatus.hidePercent) {
-        spinner.update(`Downloading ${modelName} - ${displayStatus.text}`);
-      } else {
-        spinner.update(`Downloading ${modelName} - ${displayStatus.text} - ${percent ?? 0}%`);
-      }
-    },
-  });
-  if (!result.ok) {
-    spinner.stop(result.message);
-    return false;
-  }
-  spinner.stop(`Downloaded ${modelName}`);
-  return true;
-}
-
-async function pullOllamaModelNonInteractive(
-  baseUrl: string,
-  modelName: string,
-  runtime: RuntimeEnv,
-): Promise<boolean> {
-  runtime.log(`Downloading ${modelName}...`);
-  const result = await pullOllamaModelCore({ baseUrl, modelName });
-  if (!result.ok) {
-    runtime.error(result.message);
-    return false;
-  }
-  runtime.log(`Downloaded ${modelName}`);
-  return true;
 }
 
 async function promptForOllamaCloudCredential(params: {
@@ -429,71 +200,6 @@ async function promptForOllamaCloudCredential(params: {
     credentialMode: captured.credentialMode,
     discoveryApiKey,
   };
-}
-
-function buildOllamaModelsConfig(
-  modelNames: string[],
-  discoveredModelsByName?: Map<string, OllamaModelWithContext>,
-  defaultModels: readonly OllamaCloudDefaultModel[] = [],
-) {
-  return modelNames.map((name) => {
-    const discovered = discoveredModelsByName?.get(name);
-    const defaultModel = defaultModels.find((model) => model.id === name);
-    if (defaultModel && !discovered) {
-      return buildDefaultOllamaCloudModelDefinition(defaultModel);
-    }
-    const capabilities =
-      discovered?.capabilities ?? (defaultModel ? [...defaultModel.capabilities] : undefined);
-    return buildOllamaModelDefinition(
-      name,
-      discovered?.contextWindow ?? defaultModel?.contextWindow,
-      capabilities,
-    );
-  });
-}
-
-function getOllamaLatestDedupeKey(name: string): string {
-  const normalized = normalizeLowercaseStringOrEmpty(name);
-  return normalized.endsWith(":latest") ? normalized.slice(0, -":latest".length) : normalized;
-}
-
-function isExplicitLatestOllamaModel(name: string): boolean {
-  return normalizeLowercaseStringOrEmpty(name).endsWith(":latest");
-}
-
-function shouldReplaceOllamaModelName(existing: string, candidate: string): boolean {
-  return !isExplicitLatestOllamaModel(existing) && isExplicitLatestOllamaModel(candidate);
-}
-
-function mergeUniqueModelNames(...groups: string[][]): string[] {
-  const indexByKey = new Map<string, number>();
-  const merged: string[] = [];
-  for (const group of groups) {
-    for (const name of group) {
-      const key = getOllamaLatestDedupeKey(name);
-      const existingIndex = indexByKey.get(key);
-      if (existingIndex !== undefined) {
-        const existing = expectDefined(merged[existingIndex], "indexed Ollama model name");
-        if (shouldReplaceOllamaModelName(existing, name)) {
-          merged[existingIndex] = name;
-        }
-        continue;
-      }
-      indexByKey.set(key, merged.length);
-      merged.push(name);
-    }
-  }
-  return merged;
-}
-
-function findAvailableOllamaModelName(modelName: string, availableModelNames: Iterable<string>) {
-  const wantedKey = getOllamaLatestDedupeKey(modelName);
-  for (const available of availableModelNames) {
-    if (getOllamaLatestDedupeKey(available) === wantedKey) {
-      return available;
-    }
-  }
-  return undefined;
 }
 
 function applyOllamaProviderConfig(
@@ -569,93 +275,6 @@ async function resolveHostBackedSuggestedModelNames(params: {
   return OLLAMA_SUGGESTED_MODELS_LOCAL;
 }
 
-function parseOllamaSetupShowInfo(data: {
-  model_info?: Record<string, unknown>;
-  capabilities?: unknown;
-  parameters?: unknown;
-}): Pick<OllamaModelWithContext, "contextWindow" | "capabilities"> {
-  let contextWindow: number | undefined;
-  for (const [key, value] of Object.entries(data.model_info ?? {})) {
-    if (key.endsWith(".context_length") && typeof value === "number" && value > 0) {
-      contextWindow = Math.floor(value);
-      break;
-    }
-  }
-  if (typeof data.parameters === "string") {
-    for (const line of data.parameters.split(/\r?\n/)) {
-      const match = line.trim().match(/^num_ctx\s+(-?\d+)\b/);
-      const parsed = match?.[1] ? Number.parseInt(match[1], 10) : undefined;
-      if (parsed && parsed > 0 && (contextWindow === undefined || parsed > contextWindow)) {
-        contextWindow = parsed;
-      }
-    }
-  }
-  const capabilities = Array.isArray(data.capabilities)
-    ? data.capabilities.filter((value): value is string => typeof value === "string")
-    : undefined;
-  return { contextWindow, capabilities };
-}
-
-async function inspectOllamaModelsForSetup(
-  baseUrl: string,
-  models: OllamaModelWithContext[],
-  signal?: AbortSignal,
-): Promise<{ inspected: OllamaModelWithContext[]; inspectionFailures: string[] }> {
-  const apiBase = resolveOllamaApiBase(baseUrl);
-  const inspected: OllamaModelWithContext[] = [];
-  const inspectionFailures: string[] = [];
-  for (let index = 0; index < models.length; index += OLLAMA_TOOLS_SCAN_CONCURRENCY) {
-    signal?.throwIfAborted();
-    const batch = models.slice(index, index + OLLAMA_TOOLS_SCAN_CONCURRENCY);
-    const results = await Promise.all(
-      batch.map(async (model) => {
-        try {
-          const requestSignal = signal
-            ? AbortSignal.any([AbortSignal.timeout(3000), signal])
-            : AbortSignal.timeout(3000);
-          const { response, release } = await fetchWithSsrFGuard({
-            url: `${apiBase}/api/show`,
-            init: {
-              method: "POST",
-              headers: { "Content-Type": "application/json" },
-              body: JSON.stringify({ name: model.name }),
-              signal: requestSignal,
-            },
-            signal: requestSignal,
-            policy: buildOllamaBaseUrlSsrFPolicy(apiBase),
-            auditContext: "ollama-setup.tools-scan",
-          });
-          try {
-            if (!response.ok) {
-              throw new Error(`Ollama model inspection failed with HTTP ${response.status}`);
-            }
-            const data = await readProviderJsonResponse<{
-              model_info?: Record<string, unknown>;
-              capabilities?: unknown;
-              parameters?: unknown;
-            }>(response, "ollama-setup.tools-scan");
-            return Object.assign({}, model, parseOllamaSetupShowInfo(data));
-          } finally {
-            await release();
-          }
-        } catch (error) {
-          signal?.throwIfAborted();
-          // Fail open per model: one stale/corrupt local model (observed: a
-          // months-old pull whose /api/show returns HTTP 500) must not brick
-          // the whole setup. Mark the failure with EMPTY capabilities: undefined
-          // means "never inspected" and downstream config building optimistically
-          // defaults that to supportsTools, which would advertise a broken model
-          // as tools-capable (ClawSweeper P2 on #109797).
-          inspectionFailures.push(`${model.name}: ${formatErrorMessage(error)}`);
-          return Object.assign({}, model, { capabilities: [] as string[] });
-        }
-      }),
-    );
-    inspected.push(...results);
-  }
-  return { inspected, inspectionFailures };
-}
-
 async function promptAndConfigureHostBackedOllama(params: {
   cfg: OpenClawConfig;
   mode: HostBackedOllamaInteractiveMode;
@@ -664,33 +283,24 @@ async function promptAndConfigureHostBackedOllama(params: {
   signal?: AbortSignal;
 }): Promise<OllamaSetupResult> {
   const baseUrl = await promptForOllamaBaseUrl(params.prompter, params.env);
-  const { reachable, models } = await fetchOllamaModels(baseUrl);
+  const {
+    reachable,
+    models,
+    inspectedModels,
+    discoveredModelsByName,
+    inspectionFailures,
+    hasToolsCapableModel,
+  } = await discoverOllamaModelsForSetup({
+    baseUrl,
+    inspectTools: true,
+    ...(params.signal ? { signal: params.signal } : {}),
+  });
 
   if (!reachable) {
     await params.prompter.note(buildOllamaUnreachableLines(baseUrl).join("\n"), "Ollama");
     throw new WizardCancelledError("Ollama not reachable");
   }
 
-  const firstScan = await inspectOllamaModelsForSetup(
-    baseUrl,
-    models.slice(0, OLLAMA_CONTEXT_ENRICH_LIMIT),
-    params.signal,
-  );
-  let inspectedModels = firstScan.inspected;
-  const inspectionFailures = [...firstScan.inspectionFailures];
-  const supportsTools = (model: OllamaModelWithContext) =>
-    model.capabilities?.includes("tools") === true;
-  let hasToolsCapableModel = inspectedModels.some(supportsTools);
-  if (!hasToolsCapableModel && models.length > OLLAMA_CONTEXT_ENRICH_LIMIT) {
-    const remainingScan = await inspectOllamaModelsForSetup(
-      baseUrl,
-      models.slice(OLLAMA_CONTEXT_ENRICH_LIMIT),
-      params.signal,
-    );
-    inspectedModels = [...inspectedModels, ...remainingScan.inspected];
-    inspectionFailures.push(...remainingScan.inspectionFailures);
-    hasToolsCapableModel = remainingScan.inspected.some(supportsTools);
-  }
   if (inspectionFailures.length > 0) {
     await params.prompter.note(
       [
@@ -701,7 +311,6 @@ async function promptAndConfigureHostBackedOllama(params: {
       "Ollama",
     );
   }
-  const discoveredModelsByName = new Map(inspectedModels.map((model) => [model.name, model]));
   let discoveredModelNames = models.map((model) => model.name);
   // A pull offer is only meaningful when inspection actually worked: if every
   // scan failed we cannot know what is installed, so recommending a multi-GB
@@ -834,7 +443,9 @@ export async function configureOllamaNonInteractive(params: {
   const baseUrl = resolveOllamaApiBase(
     (params.opts.customBaseUrl?.trim() || resolveOllamaSetupDefaultBaseUrl()).replace(/\/+$/, ""),
   );
-  const { reachable, models } = await fetchOllamaModels(baseUrl);
+  const { reachable, models, discoveredModelsByName } = await discoverOllamaModelsForSetup({
+    baseUrl,
+  });
   const explicitModel = normalizeOllamaModelName(params.opts.customModelId);
 
   if (!reachable) {
@@ -843,11 +454,6 @@ export async function configureOllamaNonInteractive(params: {
     return params.nextConfig;
   }
 
-  const enrichedModels = await enrichOllamaModelsWithContext(
-    baseUrl,
-    models.slice(0, OLLAMA_CONTEXT_ENRICH_LIMIT),
-  );
-  const discoveredModelsByName = new Map(enrichedModels.map((model) => [model.name, model]));
   const modelNames = models.map((model) => model.name);
   // Configured local models are advertised as available, so suggested models
   // belong in the inventory only when Ollama actually reports them as installed.
@@ -964,18 +570,3 @@ export async function ensureOllamaModelPulled(params: {
     throw new WizardCancelledError("Failed to download selected Ollama model");
   }
 }
-
-function toLintErrorObject(value: unknown, fallbackMessage: string): Error {
-  if (value instanceof Error) {
-    return value;
-  }
-  if (typeof value === "string") {
-    return new Error(value);
-  }
-  const error = new Error(fallbackMessage, { cause: value });
-  if ((typeof value === "object" && value !== null) || typeof value === "function") {
-    Object.assign(error, value);
-  }
-  return error;
-}
-/* oxlint-disable max-lines -- TODO: split this grandfathered oversized file. */
